Add test

+++ amqp_lib/decode.c
@@ -12,7 +12,7 @@ int decodeDescribedFormatCode(char *source_buffer, PDescribedFormatCode out_desc
     return 0;
 }
 
-int parseAmqp10MessageBuffer(char *source_buffer, int len, PMessage_t out_message) {
+int parseAmqp10MessageBuffer(char *source_buffer, long len, PMessage_t out_message) {
     int offset = 0;
     while (offset != len) {
         DescribedFormatCode formatCode;
