+++ amqp_lib/decode.h
@@ -5,7 +5,7 @@
 #include "types.h"
 
 
-int parseAmqp10MessageBuffer(char *source_buffer, int len, PMessage_t out_message);
+int parseAmqp10MessageBuffer(char *source_buffer, long len, PMessage_t out_message);
 
 int decodeDescribedFormatCode(char *source_buffer, PDescribedFormatCode out_described_format_code);
